Add missing cast to disambiguate pushKV.
Explicitly cast size_t values to int64_t when using pushKV in
rpc/game.cpp, so that the overload of pushKV is explicitly specified.
Otherwise there are compile errors on Win32 and Mac.

+++ src/rpc/game.cpp
@@ -291,8 +291,8 @@ game_sendupdates (const JSONRPCRequest& request)
   result.pushKV ("ancestor", ancestor->GetBlockHash ().GetHex ());
   result.pushKV ("reqtoken", reqtoken);
   UniValue steps(UniValue::VOBJ);
-  steps.pushKV ("detach", w.detach.size ());
-  steps.pushKV ("attach", w.attach.size ());
+  steps.pushKV ("detach", static_cast<int64_t> (w.detach.size ()));
+  steps.pushKV ("attach", static_cast<int64_t> (w.attach.size ()));
   result.pushKV ("steps", steps);
 
   GetGameBlocksNotifier ();
